reduxified the toggling of the intro component, cleaned up some styling, and modularized the synth voice function
