+++ .gitignore
@@ -5,3 +5,4 @@ sync-request.json
 Zeiza-6f0fd6f67875.json
 Zeiza-6464ec06fd66.json
 in.json
+.gitignore
