+++ config/zeizaCommands.js
@@ -8,8 +8,8 @@ var knownCommands = [
             'make background colour',
             'update background colour',
             'change the colour of the background',
-            "make the background's color",
-            "change the background's color",
+            "make the background's colour",
+            "change the background's colour",
             "update the background's colour",
             'update the background colour',
             'change the background colour',
