+++ src/actions.js
@@ -53,3 +53,12 @@ export const displayMostRecentAIResponse = zeizaResponse => {
 ////////////////////////////////////////////////////////////////////////////////
 ////////////////////////// END PUSH USER / AI COMMENTS TO STATE ////////////////
 ////////////////////////////////////////////////////////////////////////////////
+
+
+
+
+
+
+export const toggleShowIntroComponent = showIntroComponent => {
+    return { type: "TOGGLE_INTRO_COMPONENT", showIntroComponent };
+}
